# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/teacher": {
      "post": {
        "operationId": "UserController_createTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/coordinator": {
      "post": {
        "operationId": "UserController_createCoordinator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCoordinatorReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/student": {
      "post": {
        "operationId": "UserController_createStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/viewer": {
      "post": {
        "operationId": "UserController_createViewer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UserController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "COORDINATOR",
                  "TEACHER",
                  "STUDENT",
                  "SYSADMIN",
                  "VIEWER"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/coordinator/{id}": {
      "delete": {
        "operationId": "UserController_deleteCoordinator",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UserController_updateCoordinatorById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateByIdBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/teacher/{id}": {
      "delete": {
        "operationId": "UserController_deleteTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UserController_updateTeacherById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateByIdBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/student/{id}": {
      "delete": {
        "operationId": "UserController_deleteStudent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UserController_updateStudentById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateByIdBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/viewer/{id}": {
      "delete": {
        "operationId": "UserController_deleteViewer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UserController_updateViewerById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateByIdBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/password": {
      "patch": {
        "operationId": "UserController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{userId}/role": {
      "patch": {
        "operationId": "UserController_changeUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleBodyReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIPPIE",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserReqDto": {
        "type": "object",
        "properties": {
          "registration": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          }
        },
        "required": [
          "registration",
          "name",
          "password"
        ]
      },
      "UserResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "registration": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "registration",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserCoordinatorReqDto": {
        "type": "object",
        "properties": {
          "registration": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          }
        },
        "required": [
          "registration",
          "name",
          "password",
          "courseId"
        ]
      },
      "MetadataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "itemsPerPage",
          "totalPages",
          "totalItems"
        ]
      },
      "PaginationResDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "required": [
          "items",
          "metadata"
        ]
      },
      "UpdateByIdBodyReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          }
        }
      },
      "ChangePasswordBodyReqDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "currentPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "currentPassword"
        ]
      },
      "ChangeRoleBodyReqDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string"
          },
          "userRole": {
            "type": "string",
            "enum": [
              "COORDINATOR",
              "TEACHER"
            ]
          }
        },
        "required": [
          "courseId",
          "userRole"
        ]
      },
...
    }
  }
}